refactor: Add typescript support
