+++ package.json
@@ -12,15 +12,15 @@
   "bugs": {
     "url": "https://github.com/nextcloud/nextcloud-cookbook/issues"
   },
-  "main": "src/main.js",
+  "main": "src/main.ts",
   "scripts": {
     "build": "npx webpack --node-env production --progress --config webpack.config.js",
     "build-bundle-analyzer": "npx webpack --node-env development --progress --config webpack.devel.js --env BUNDLE_ANALYZER=true",
     "build-dev": "npx webpack --node-env development --progress --config webpack.devel.js",
     "ci-test": "NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" npx jest --ci",
     "dev": "npx webpack --node-env development --progress --watch --config webpack.devel.js",
-    "eslint": "npx eslint --cache --cache-strategy content 'src/**/*.{vue,js}'",
-    "eslint-fix": "npx eslint --cache --cache-strategy content --fix 'src/**/*.{vue,js}'",
+    "eslint": "npx eslint --cache --cache-strategy content 'src/**/*.{vue,js,ts}'",
+    "eslint-fix": "npx eslint --cache --cache-strategy content --fix 'src/**/*.{vue,js,ts}'",
     "package-lint": "npx prettier-package-json --write ./package.json",
     "prettier": "npx prettier --check src",
     "prettier-fix": "npx prettier --write src",
@@ -57,14 +57,20 @@
     "@nextcloud/browserslist-config": "^3.0.0",
     "@nextcloud/stylelint-config": "^2.3.1",
     "@nextcloud/webpack-vue-config": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.18.0",
+    "@typescript-eslint/parser": "^6.18.0",
+    "@vue/cli-plugin-typescript": "~5.0.8",
+    "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/vue2-jest": "^29.2.6",
     "babel-jest": "^29.7.0",
     "check-peer-dependencies": "^4.3.0",
     "clean-webpack-plugin": "^4.0.0",
     "eslint": "^8.53.0",
     "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-vue": "^9.18.1",
     "jest": "^29.7.0",
@@ -73,6 +79,8 @@
     "stylelint": "^15.11.0",
     "stylelint-config-idiomatic-order": "^10.0.0",
     "stylelint-scss": "^5.3.0",
+    "ts-loader": "^9.5.1",
+    "typescript": "~5.3.3",
     "vue-template-compiler": "^2.7.15",
     "webpack": "^5.89.0",
     "webpack-bundle-analyzer": "^4.9.1",
