+++ src/js/helper.js
@@ -74,7 +74,12 @@ function isSameItemInstance(url1, url2) {
     return comps1[2] === comps2[2];
 }
 
-// A simple function to sanitize HTML tags
+/**
+ * A simple function to sanitize HTML tags.
+ * @param {string} text Input string
+ * @returns {string}
+ */
+
 function escapeHTML(text) {
     return text.replace(
         /["&'<>]/g,
@@ -143,7 +148,7 @@ function notify(title, options) {
         return;
     }
     if (Notification.permission === 'granted') {
-        // eslint-disable-next-line no-unused-vars
+        // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
         const notification = new Notification(title, options);
     } else if (Notification.permission !== 'denied') {
         Notification.requestPermission((permission) => {
@@ -151,7 +156,7 @@ function notify(title, options) {
                 Notification.permission = permission;
             }
             if (permission === 'granted') {
-                // eslint-disable-next-line no-unused-vars
+                // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
                 const notification = new Notification(title, options);
             } else {
                 showSimpleAlertModal(title);
